apps/api/requirements.txt
# Omniparse AI Stack - API Dependencies
# Python 3.10+

# Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Async Task
celery[redis]>=5.3.0
redis>=5.0.0

# Config & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# OCR
easyocr>=1.7.0
# paddleocr>=2.7.0  # 선택적: PaddleOCR 사용 시 주석 해제
# paddlepaddle>=2.5.0

# Parser (기본: 휴리스틱, 향후 Unstructured/LayoutParser 확장 가능)
# unstructured>=0.11.0
# layoutparser>=0.3.4

# Image (PIL 기본, 향후 PyTorch/YOLO 확장)
Pillow>=10.0.0

# Agent (LangChain 기반 Q&A)
langchain>=0.2.0
langchain-openai>=0.0.2
langchain-core>=0.2.0

# Database (향후 확장)
# sqlalchemy[asyncio]>=2.0.0
# asyncpg>=0.29.0
